Fix ApplyButton route and properly rename ButtonComponents folder
